refactor(38d-T5b): 搬移 test_new_scrapers.py mock-only 測試到 unit 層
- integration test_new_scrapers.py: 92 → 6 tests（僅保留 TestClient）
- 81 mock-only tests 搬到 9 個 unit files（3 合併 + 6 新建）
- 刪除 5 個內部重複（TestFastPathRouting + TestDMMProgressiveFacade）
- unit: 1030 → 1111, integration: 296 → 210
