+++ tests/unit/test_scraper_dmm_direct.py
@@ -3,8 +3,12 @@ Tests for Phase 37d T3 — Proxy `direct` 模式
 
 覆蓋 _is_dmm_enabled() / _dmm_proxy_url() helpers 和 DMMScraper 行為。
 """
+import json
 import pytest
 from unittest.mock import patch, MagicMock
+
+from core.scrapers.dmm import DMMScraper
+from core.scrapers.models import ScraperConfig
 
 
 # ── TestIsDmmEnabled ──────────────────────────────────────────────────────────
@@ -122,3 +126,124 @@ class TestSearchDirect:
         source = _get_fuzzy_source('dmm', proxy_url)
         assert source == 'javbus', \
             "proxy_url='' + primary_source='dmm' 時必須 fallback 到 javbus"
+
+
+# ============================================================
+# Mock Data (from test_new_scrapers.py)
+# ============================================================
+
+DMM_SEARCH_RESPONSE = {
+    "data": {
+        "legacySearchPPV": {
+            "result": {
+                "contents": [{"id": "sone00205"}]
+            }
+        }
+    }
+}
+
+DMM_DETAIL_RESPONSE = {
+    "data": {
+        "ppvContent": {
+            "id": "sone00205",
+            "title": "成人への卒業",
+            "description": "テスト",
+            "packageImage": {"largeUrl": "https://pics.dmm.co.jp/sone205pl.jpg"},
+            "makerReleasedAt": "2024-03-19T00:00:00+09:00",
+            "duration": 120,
+            "actresses": [{"name": "Nana Miho"}],
+            "directors": [],
+            "series": {"name": ""},
+            "maker": {"name": "S1 NO.1 STYLE"},
+            "makerContentId": "SONE-205",
+        }
+    }
+}
+
+
+def _make_mock_resp(status_code=200, json_data=None, content=None):
+    """Build a MagicMock that mimics requests.Response."""
+    mock_resp = MagicMock()
+    mock_resp.status_code = status_code
+    if json_data is not None:
+        mock_resp.json = lambda: json_data
+    if content is not None:
+        mock_resp.content = content
+    return mock_resp
+
+
+# ============================================================
+# Tests merged from integration/test_new_scrapers.py TestDMMScraper
+# ============================================================
+
+class TestDMMScraperIntegration:
+    """DMM scraper tests (merged from test_new_scrapers.py)"""
+
+    @pytest.fixture
+    def dmm_scraper(self, tmp_path, monkeypatch):
+        """DMM scraper with isolated cache files"""
+        import core.scrapers.dmm as dmm_module
+        monkeypatch.setattr(dmm_module, "CACHE_FILE", tmp_path / "dmm_content_ids.json")
+        monkeypatch.setattr(dmm_module, "PREFIX_FILE", tmp_path / "dmm_prefix_hints.json")
+        config = ScraperConfig(proxy_url="http://test-proxy:8080")
+        return DMMScraper(config)
+
+    def test_dmm_no_proxy_session_proxies_not_set(self):
+        """無 proxy_url 時 session.proxies 不被設定（直連模式）"""
+        scraper = DMMScraper()  # 無 proxy_url → 直連模式
+        assert not scraper._session.proxies, \
+            "proxy_url='' 時 session.proxies 不應被設定"
+
+    def test_dmm_cache_hit(self, dmm_scraper, tmp_path, monkeypatch):
+        """快取命中時不呼叫 search query（detail query + probe query，不超過 2 次）"""
+        import core.scrapers.dmm as dmm_module
+        cache_path = tmp_path / "dmm_content_ids.json"
+        cache_path.write_text('{"SONE-205": "sone00205"}', encoding='utf-8')
+
+        detail_resp = _make_mock_resp(status_code=200, json_data=DMM_DETAIL_RESPONSE)
+
+        with patch.object(dmm_scraper._session, 'post', return_value=detail_resp) as mock_post:
+            with patch('core.scrapers.utils.rate_limit'):
+                video = dmm_scraper.search("SONE-205")
+
+        assert video is not None
+        assert video.title == "成人への卒業"
+        assert video.number == "SONE-205"
+        for call_args in mock_post.call_args_list:
+            payload = call_args[1].get('json', {}) if call_args[1] else {}
+            query_str = payload.get('query', '')
+            assert 'legacySearchPPV' not in query_str, "Cache hit should not trigger search query"
+
+    def test_dmm_graphql_success(self, dmm_scraper):
+        """無快取時依次呼叫 search query + detail query，成功返回 Video"""
+        search_resp = _make_mock_resp(status_code=200, json_data=DMM_SEARCH_RESPONSE)
+        detail_resp = _make_mock_resp(status_code=200, json_data=DMM_DETAIL_RESPONSE)
+
+        with patch.object(dmm_scraper._session, 'post', side_effect=[
+            _make_mock_resp(status_code=404),  # _convert_with_hints → _fetch_by_id → 404
+            search_resp,                        # _search_content_id
+            detail_resp,                        # _fetch_by_id(discovered_cid)
+        ]):
+            with patch('core.scrapers.utils.rate_limit'):
+                video = dmm_scraper.search("SONE-205")
+
+        assert video is not None
+        assert video.number == "SONE-205"
+        assert video.title == "成人への卒業"
+        assert video.source == "dmm"
+        assert "dmm.co.jp" in video.detail_url
+        assert video.date == "2024-03-19"
+        assert len(video.actresses) == 1
+        assert video.actresses[0].name == "Nana Miho"
+        assert video.maker == "S1 NO.1 STYLE"
+
+    def test_dmm_cache_isolation(self, dmm_scraper, tmp_path):
+        """搜尋成功後 cache 寫入 tmp_path，不污染 project root"""
+        detail_resp = _make_mock_resp(status_code=200, json_data=DMM_DETAIL_RESPONSE)
+
+        with patch.object(dmm_scraper._session, 'post', return_value=detail_resp):
+            with patch('core.scrapers.utils.rate_limit'):
+                video = dmm_scraper.search("SONE-205")
+
+        # cache 應寫入 tmp_path 而非 project root
+        assert (tmp_path / "dmm_content_ids.json").exists()
